backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Scheduler
apscheduler>=3.10.0

# LangChain/LangGraph (simplified - only for pipeline orchestration)
langgraph>=0.2.0

# AI - OpenAI for LLM-based scoring
openai>=1.10.0

# Job Scraping - JobSpy (Real-time job scraping)
python-jobspy>=1.1.75

# Document Processing
pypdf2>=3.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Configuration
python-dotenv>=1.0.0
